Refactored create/edit assignment WiP, fixed add yjs assignment structure
