+++ src/lib/planning/createPlanningDocument.ts
@@ -1,0 +1,131 @@
+import * as Y from 'yjs'
+import { newsDocToYMap } from '../../../src-srv/utils/transformations/yjs/yMap'
+import { Document } from '@/protos/service'
+
+/**
+ * Create empty planning document and convert it to Y.Doc as appropriate.
+ * Returns array with uuid and Y.Doc.
+ *
+ * @returns [string, Y.Doc]
+ */
+export const createPlanningDocument = (): [string, Y.Doc] => {
+  const documentId = crypto.randomUUID()
+  const yDoc = new Y.Doc()
+  const yEle = yDoc.getMap('ele')
+
+  newsDocToYMap(getPlanningTemplate(documentId), yEle)
+
+  return [documentId, yDoc]
+}
+
+/**
+ * Create a template structure for a aplanning
+ *
+ * TODO: Should be refactored into a more coherent group of functions with createPlanningAssignment etc
+ */
+function getPlanningTemplate(id: string): Document {
+  return {
+    // @ts-expect-error TODO: Remove this line when merged with snapshot-functionality branch
+    __inPogress: true,
+    uuid: id,
+    type: 'core/planning-item',
+    uri: `core://newscoverage/${id}`,
+    url: '',
+    title: '',
+    language: 'sv-se',
+    content: [],
+    meta: [
+      {
+        id: '',
+        uuid: '',
+        uri: '',
+        url: '',
+        type: 'core/planning-item',
+        title: '',
+        data: {
+          public: 'true',
+          end_date: '',
+          tentative: 'false',
+          start_date: ''
+        },
+        rel: '',
+        role: '',
+        name: '',
+        value: '',
+        contentType: '',
+        links: [],
+        content: [],
+        meta: []
+      },
+      {
+        id: '',
+        uuid: '',
+        uri: '',
+        url: '',
+        type: 'core/newsvalue',
+        title: '',
+        data: {},
+        rel: '',
+        role: '',
+        name: '',
+        value: '',
+        contentType: '',
+        links: [],
+        content: [],
+        meta: []
+      },
+      {
+        id: '',
+        uuid: '',
+        uri: '',
+        url: '',
+        type: 'tt/slugline',
+        title: '',
+        data: {},
+        rel: '',
+        role: '',
+        name: '',
+        value: '',
+        contentType: '',
+        links: [],
+        content: [],
+        meta: []
+      },
+      {
+        id: '',
+        uuid: '',
+        uri: '',
+        url: '',
+        type: 'core/description',
+        title: '',
+        data: { text: '' },
+        rel: '',
+        role: 'public',
+        name: '',
+        value: '',
+        contentType: '',
+        links: [],
+        content: [],
+        meta: []
+      },
+      {
+        id: '',
+        uuid: '',
+        uri: '',
+        url: '',
+        type: 'core/description',
+        title: '',
+        data: { text: '' },
+        rel: '',
+        role: 'internal',
+        name: '',
+        value: '',
+        contentType: '',
+        links: [],
+        content: [],
+        meta: []
+      }
+    ],
+    links: []
+  }
+}
