+++ src/views/Planning/components/index.tsx
@@ -1,4 +1,4 @@
-export { Assignments as PlanAssignments } from './Assignments'
+export { AssignmentTable } from './AssignmentTable'
 export { PlanDate } from './PlanDate'
 export { PlanDescription } from './PlanDescription'
 export { PlanNewsvalue } from './PlanNewsvalue'
