+++ src/views/PlanningOverview/PlanningHeader/CreatePlan.tsx
@@ -7,7 +7,7 @@ import * as Y from 'yjs'
 import { Planning } from '@/views/Planning'
 import { PlusIcon } from '@ttab/elephant-ui/icons'
 import { useState } from 'react'
-import { newsDocToYMap } from '../../../../src-srv/utils/transformations/yjs/yMap'
+import { createPlanningDocument } from '@/lib/planning/createPlanningDocument'
 
 export const CreatePlan = (): JSX.Element => {
   const [planning, setPlanning] = useState<[string | undefined, Y.Doc | undefined]>([undefined, undefined])
@@ -49,121 +49,4 @@ export const CreatePlan = (): JSX.Element => {
       </DialogContent>
     </Dialog>
   )
-}
-
-/**
- * Create empty planning document and convert it to Y.Doc as appropriate. Returns
- * array with uuid and Y.Doc.
- *
- * @returns [string, Y.Doc]
- */
-const createPlanningDocument = (): [string, Y.Doc] => {
-  const documentId = crypto.randomUUID()
-  const yDoc = new Y.Doc()
-  const yEle = yDoc.getMap('ele')
-
-  newsDocToYMap({
-    uuid: documentId,
-    type: 'core/planning-item',
-    uri: `core://newscoverage/${documentId}`,
-    url: '',
-    title: '',
-    language: 'sv-se',
-    content: [],
-    meta: [
-      {
-        id: '',
-        uuid: '',
-        uri: '',
-        url: '',
-        type: 'core/planning-item',
-        title: '',
-        data: {
-          public: 'true',
-          end_date: '',
-          tentative: 'false',
-          start_date: ''
-        },
-        rel: '',
-        role: '',
-        name: '',
-        value: '',
-        contentType: '',
-        links: [],
-        content: [],
-        meta: []
-      },
-      {
-        id: '',
-        uuid: '',
-        uri: '',
-        url: '',
-        type: 'core/newsvalue',
-        title: '',
-        data: {},
-        rel: '',
-        role: '',
-        name: '',
-        value: '',
-        contentType: '',
-        links: [],
-        content: [],
-        meta: []
-      },
-      {
-        id: '',
-        uuid: '',
-        uri: '',
-        url: '',
-        type: 'tt/slugline',
-        title: '',
-        data: {},
-        rel: '',
-        role: '',
-        name: '',
-        value: '',
-        contentType: '',
-        links: [],
-        content: [],
-        meta: []
-      },
-      {
-        id: '',
-        uuid: '',
-        uri: '',
-        url: '',
-        type: 'core/description',
-        title: '',
-        data: { text: '' },
-        rel: '',
-        role: 'public',
-        name: '',
-        value: '',
-        contentType: '',
-        links: [],
-        content: [],
-        meta: []
-      },
-      {
-        id: '',
-        uuid: '',
-        uri: '',
-        url: '',
-        type: 'core/description',
-        title: '',
-        data: { text: '' },
-        rel: '',
-        role: 'internal',
-        name: '',
-        value: '',
-        contentType: '',
-        links: [],
-        content: [],
-        meta: []
-      }
-    ],
-    links: []
-  }, yEle)
-
-  return [documentId, yDoc]
 }
